chore: move mock data to http
